refactor(readme): generate instrumentation from schema
