+++ Cargo.toml
@@ -37,10 +37,6 @@ members = [
     "domains/query_engine/model",
     "domains/query_engine/server",
     "domains/query_engine/ui",
-    # C++ integration example (uses readme model)
-    "examples/cpp-integration/bridge",
-    # Python integration example (uses readme model)
-    "examples/python-integration/bridge",
     # Query engine C++ bridge test
     "domains/query_engine/tests/cpp/instrumentation",
     "domains/query_engine/tests/cpp/bridge",
@@ -49,6 +45,12 @@ members = [
     "domains/query_engine/tests/python/bridge",
     # README example
     "examples/readme",
+    # Legacy README example using model macros
+    "examples/legacy/readme",
+    # Legacy C++ integration example using the legacy README model
+    "examples/legacy/cpp-integration/bridge",
+    # Legacy Python integration example using the legacy README model
+    "examples/legacy/python-integration/bridge",
     # Simulator crates
     "examples/simulator/analyzer",
     "examples/simulator/application",
@@ -107,7 +109,7 @@ default-members = [
     "domains/query_engine/model",
     "domains/query_engine/server",
     "domains/query_engine/ui",
-    "examples/cpp-integration/bridge",
+    "examples/legacy/cpp-integration/bridge",
     "domains/query_engine/tests/cpp/instrumentation",
     "domains/query_engine/tests/cpp/bridge",
     "examples/readme",
