+++ examples/legacy/cpp-integration/bridge/.gitignore
