+++ examples/legacy/cpp-integration/bridge/src/lib.rs
