+++ examples/legacy/cpp-integration/cpp/.gitignore
