+++ examples/legacy/cpp-integration/cpp/src/main.cpp
