+++ examples/legacy/readme/.gitignore
@@ -1,0 +1,1 @@
+events/
