+++ examples/legacy/readme/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "quent-legacy-readme-example"
+version.workspace = true
+edition.workspace = true
+publish.workspace = true
+
+[features]
+collector = ["quent-model/collector"]
+
+[dependencies]
+quent-model = { path = "../../../crates/model" }
+serde.workspace = true
+uuid = { workspace = true, features = ["serde", "v7"] }
