+++ examples/readme/build.rs
@@ -1,0 +1,27 @@
+// SPDX-FileCopyrightText: Copyright (c) 2026, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+use std::path::Path;
+
+use quent_instrumentation_build::{Options, generate};
+
+fn main() -> Result<(), Box<dyn std::error::Error>> {
+    let model = Path::new(env!("CARGO_MANIFEST_DIR")).join("model.yaml");
+    println!("cargo:rerun-if-changed=build.rs");
+    println!("cargo:rerun-if-changed={}", model.display());
+
+    let parsed = quent_yaml::parse_from_file(&model)?;
+    for warning in &parsed.warnings {
+        println!("cargo:warning={warning}");
+    }
+
+    generate(
+        &parsed.schema,
+        &Options {
+            serde: true,
+            ..Options::default()
+        },
+    )?;
+
+    Ok(())
+}
